# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reddit Content Platform API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Reddit Content Platform",
      "url": "https://github.com/your-username/reddit-content-platform",
      "email": "support@yourcompany.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-api-version": "1.0.0",
    "x-api-versions-supported": [
      "v1"
    ],
    "x-api-version-header": "X-API-Version",
    "x-api-media-type": "application/vnd.reddit-platform.v1+json"
  },
  "paths": {
...
    "/api/v1/system/alerts": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Alerts",
        "description": "Get system alerts and warnings.\n\nArgs:\n    resolved: Filter by resolved status (None for all)\n    current_user: Authenticated user\n    \nReturns:\n    List of system alerts\n    \nRaises:\n    HTTPException: If alert retrieval fails",
        "operationId": "get_system_alerts_api_v1_system_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolved status",
              "title": "Resolved"
            },
            "description": "Filter by resolved status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertInfo"
                  },
                  "title": "Response Get System Alerts Api V1 System Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/info": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get System Info",
        "description": "Get system information and configuration.\n\nArgs:\n    current_user: Authenticated user\n    \nReturns:\n    System information\n    \nRaises:\n    HTTPException: If system info retrieval fails",
        "operationId": "get_system_info_api_v1_system_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get System Info Api V1 System Info Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "message",
          "timestamp"
        ],
        "title": "AlertInfo",
        "description": "Alert information model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://www.reddit.com/api/v1/authorize",
            "tokenUrl": "/api/v1/auth/token",
            "scopes": {
              "read": "Read Reddit posts and comments",
              "identity": "Access user identity information"
            }
          }
        },
        "description": "Reddit OAuth2 authentication"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-app.railway.app",
      "description": "Production server"
    }
  ],
...
}